pyproject.toml
[project]
name = "nova"
version = "0.1.0"
description = "Personal home assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.1.1",
    "pyaudio>=0.2.14",
    "openwakeword>=0.4.0",
]

[dependency-groups]
dev = ["pyaudio>=0.2.14", "types-pyaudio>=0.2.16.20250801"]
